Replace ScreenName annotation with custom implementation.
Since the ScreenName annotation is intended to be used for user-displayable
information and that may not be desirable as the logged name in the logger,
or particularly in analytics, the annotation has been removed and
replaced with a custom implementation.
Since that implementation did not require i18n, we can accept a plain
string as the value, eliminating our need to require android Resources
as a dependency as well as Prism . This change causes a BC break.